+++ src/core.c
@@ -16,6 +16,7 @@
 
 // nanonext - C level - Core Functions -----------------------------------------
 
+#define NANONEXT_SIGNALS
 #include "nanonext.h"
 
 // internals -------------------------------------------------------------------
@@ -122,6 +123,32 @@ static SEXP nano_outHook(SEXP x, SEXP fun) {
 
 // functions with forward definitions in nanonext.h ----------------------------
 
+void raio_complete_interrupt(void *arg) {
+
+  nano_aio *raio = (nano_aio *) arg;
+  int res = nng_aio_result(raio->aio);
+  if (res == 0) {
+    nng_msg *msg = nng_aio_get_msg(raio->aio);
+    raio->data = msg;
+    nng_pipe p = nng_msg_get_pipe(msg);
+    res = - (int) p.id;
+  }
+
+  raio->result = res;
+
+  if (raio->cb != NULL)
+    later2(raio_invoke_cb, raio->cb);
+
+  if (nano_interrupt) {
+#ifdef _WIN32
+    UserBreak = 1;
+#else
+    kill(getpid(), SIGINT);
+#endif
+  }
+
+}
+
 void raio_complete_signal(void *arg) {
 
   nano_aio *raio = (nano_aio *) arg;
@@ -129,12 +156,16 @@ void raio_complete_signal(void *arg) {
   nng_cv *cv = ncv->cv;
   nng_mtx *mtx = ncv->mtx;
 
-  const int res = nng_aio_result(raio->aio);
-  if (res == 0)
-    raio->data = nng_aio_get_msg(raio->aio);
+  int res = nng_aio_result(raio->aio);
+  if (res == 0) {
+    nng_msg *msg = nng_aio_get_msg(raio->aio);
+    raio->data = msg;
+    nng_pipe p = nng_msg_get_pipe(msg);
+    res = - (int) p.id;
+  }
 
   nng_mtx_lock(mtx);
-  raio->result = res - !res;
+  raio->result = res;
   ncv->condition++;
   nng_cv_wake(cv);
   nng_mtx_unlock(mtx);